// src/app/module/notices/model/notice.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Company } from './company';
import { NoticePriority } from './noticePriority';


export interface Notice {
    company?: Company;
    description?: string;
    id?: number;
    priority?: NoticePriority;
    validate?: string;
}