// src/app/module/notices/model/noticePriority.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Company } from './company';


export interface NoticePriority {
    color?: string;
    company?: Company;
    id?: number;
    level?: number;
    name?: string;
}